+++ backend/config.py
@@ -1,0 +1,10 @@
+import os
+
+class Config(object):
+    DEBUG = True
+    SECRET_KEY = os.environ.get('SECRET_KEY', 'default-secret-key')
+    base_dir = os.path.dirname(__file__)
+    database = 'sqlite:///' + os.path.join(base_dir, 'data.sqlite')
+    SQLALCHEMY_DATABASE_URI = 'sqlite:///my_database.db'
+    SQLALCHEMY_DATABASE_URI = database
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
