+++ frontend/src/components/Login.jsx
@@ -7,7 +7,7 @@ const Login = ({ setAccessToken }) => {
 
     const SignIn = async (e) => {
         e.preventDefault()
-        const res = await fetch('http://localhost:5000/login', {
+        const res = await fetch('http://localhost:5000/auth/login', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json',
