+++ frontend/src/components/Register.jsx
@@ -8,7 +8,7 @@ const Resister = () => {
 
     const SignUp = async (e) => {
         e.preventDefault()
-        const res = await fetch('http://localhost:5000/register', {
+        const res = await fetch('http://localhost:5000/auth/register', {
         method: 'POST',
         headers: {
             'Content-Type': 'application/json',
